Split retrieve.util into whois and URLVoid
